# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DC API
  version: 1.0.0
  description: 数据连接器 API 契约（MVP）
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: 查询任务列表
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksResponse"
    post:
      tags: [tasks]
      summary: 创建任务实例
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/v1/tasks/{id}:
    patch:
      tags: [tasks]
      summary: 更新任务实例（参数/存储/触发）
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/v1/tasks/{id}/run:
    post:
      tags: [tasks]
      summary: 手动触发任务执行
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunTriggerResponse"
  /api/v1/tasks/{id}/retry:
    post:
      tags: [tasks]
      summary: 重试最近一次失败执行
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunTriggerResponse"
  /api/v1/tasks/{id}/sidebar:
    get:
      tags: [tasks]
      summary: 获取任务侧边栏联动数据（参数/存储/日志）
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSidebarResponse"
  /api/v1/tasks/{id}/sidebar/params:
    patch:
      tags: [tasks]
      summary: 保存任务侧边栏参数配置
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskSidebarParamsRequest"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/v1/tasks/{id}/sidebar/storage:
    patch:
      tags: [tasks]
      summary: 保存任务侧边栏存储配置
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskSidebarStorageRequest"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/v1/tasks/{id}/sidebar/logs:
    get:
      tags: [tasks]
      summary: 查询任务侧边栏指定执行批次日志
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: run_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunLogsResponse"
...
components:
  schemas:
    BaseResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
...
    Task:
      type: object
      required: [id, app_id, account_id, storage_config_id, notification_config_id, cron_expr, status]
      properties:
        id:
          type: integer
        app_id:
          type: integer
        account_id:
          type: integer
        storage_config_id:
          type: integer
        notification_config_id:
          type: integer
        cron_expr:
          type: string
        status:
          type: string
          enum: [enabled, paused]
        name:
          type: string
        app_name:
          type: string
        platform_code:
          type: string
        platform_name:
          type: string
        shop_name:
          type: string
        last_run_at:
          type: string
          format: date-time
          nullable: true
        last_run_status:
          type: string
          nullable: true
          description: 最近一次执行状态，来源 task_run.status
          enum: [pending, running, success, failed, cancelled]
        params:
          type: object
          additionalProperties: true
...
    CreateTaskRequest:
      type: object
      required: [app_id, account_id, name, cron_expr, storage_config_id, notification_config_id]
      properties:
        app_id:
          type: integer
        account_id:
          type: integer
        name:
          type: string
        cron_expr:
          type: string
        params:
          type: object
          additionalProperties: true
        storage_config_id:
          type: integer
        notification_config_id:
          type: integer
    UpdateTaskRequest:
      type: object
      properties:
        name:
          type: string
        cron_expr:
          type: string
        status:
          type: string
          enum: [enabled, paused]
        storage_config_id:
          type: integer
        notification_config_id:
          type: integer
        params:
          type: object
          additionalProperties: true
    UpdateTaskSidebarParamsRequest:
      type: object
      required: [default_download_days]
      properties:
        default_download_days:
          type: integer
          minimum: 1
        extra_params:
          type: object
          additionalProperties: true
    UpdateTaskSidebarStorageRequest:
      type: object
      required: [storage_config_id]
      properties:
        storage_config_id:
          type: integer
...
    TaskResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Task"
    TasksResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Task"
    RunTriggerResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              required: [task_id, run_id, queued]
              properties:
                task_id:
                  type: integer
                run_id:
                  type: integer
                queued:
                  type: boolean
...
    RunLogsResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                run_id:
                  type: integer
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      step:
                        type: string
                      level:
                        type: string
                        enum: [INFO, WARN, ERROR]
                      message:
                        type: string
                      ext:
                        type: object
                        additionalProperties: true
                      ts:
                        type: string
                        format: date-time
    TaskSidebarResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                task_id:
                  type: integer
                params:
                  type: object
                  properties:
                    default_download_days:
                      type: integer
                    extra_params:
                      type: object
                      additionalProperties: true
                storage:
                  type: object
                  properties:
                    current_storage_id:
                      type: integer
                    current_storage:
                      type: object
                      nullable: true
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        type:
                          type: string
                        status:
                          type: string
                        config:
                          type: object
                          additionalProperties: true
                    options:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          type:
                            type: string
                          status:
                            type: string
                logs:
                  type: object
                  properties:
                    runs:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          status:
                            type: string
                          started_at:
                            type: string
                            nullable: true
                          ended_at:
                            type: string
                            nullable: true
                          duration_ms:
                            type: integer
                            nullable: true
                          error_code:
                            type: string
                            nullable: true
                          error_message:
                            type: string
                            nullable: true
                    selected_run_id:
                      type: integer
                      nullable: true
                    items:
                      type: array
                      items:
                        type: object
                        properties:
                          step:
                            type: string
                          level:
                            type: string
                          message:
                            type: string
                          ext:
                            type: object
                            additionalProperties: true
                          ts:
                            type: string
                            nullable: true
...